Разделил логику и работу с базой в моделях. При увеличении событий будет проще поддерживать.
