+++ models/History.php
@@ -4,34 +4,11 @@ namespace app\models;
 
 use app\models\traits\ObjectNameTrait;
 use Yii;
-use app\models\events\Customer;
-use app\models\events\User;
-use yii\db\ActiveQuery;
-use yii\db\ActiveRecord;
+
 use app\models\interfaces\HistoryInterface;
-/**
- * This is the model class for table "{{%history}}".
- *
- * @property integer $id
- * @property string $ins_ts
- * @property integer $customer_id
- * @property string $event
- * @property string $object
- * @property integer $object_id
- * @property string $message
- * @property string $detail
- * @property integer $user_id
- *
- * @property string $eventText
- *
- * @property Customer $customer
- * @property User $user
- *
- * @property Task $task
- * @property Sms $sms
- * @property Call $call
- */
-class History extends ActiveRecord implements HistoryInterface
+use app\models\repositories\History  as HistoryActiveRecord;
+
+class History extends HistoryActiveRecord implements HistoryInterface
 {
     use ObjectNameTrait;
 
@@ -50,64 +27,6 @@ class History extends ActiveRecord implements HistoryInterface
 
     const EVENT_CUSTOMER_CHANGE_TYPE = 'customer_change_type';
     const EVENT_CUSTOMER_CHANGE_QUALITY = 'customer_change_quality';
-
-    /**
-     * @inheritdoc
-     */
-    public static function tableName()
-    {
-        return '{{%history}}';
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function rules()
-    {
-        return [
-            [['ins_ts'], 'safe'],
-            [['customer_id', 'object_id', 'user_id'], 'integer'],
-            [['event'], 'required'],
-            [['message', 'detail'], 'string'],
-            [['event', 'object'], 'string', 'max' => 255],
-            [['customer_id'], 'exist', 'skipOnError' => true, 'targetClass' => Customer::class, 'targetAttribute' => ['customer_id' => 'id']],
-            [['user_id'], 'exist', 'skipOnError' => true, 'targetClass' => User::class, 'targetAttribute' => ['user_id' => 'id']],
-        ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function attributeLabels()
-    {
-        return [
-            'id' => Yii::t('app', 'ID'),
-            'ins_ts' => Yii::t('app', 'Ins Ts'),
-            'customer_id' => Yii::t('app', 'Customer ID'),
-            'event' => Yii::t('app', 'Event'),
-            'object' => Yii::t('app', 'Object'),
-            'object_id' => Yii::t('app', 'Object ID'),
-            'message' => Yii::t('app', 'Message'),
-            'detail' => Yii::t('app', 'Detail'),
-            'user_id' => Yii::t('app', 'User ID'),
-        ];
-    }
-
-    /**
-     * @return ActiveQuery
-     */
-    public function getCustomer()
-    {
-        return $this->hasOne(Customer::class, ['id' => 'customer_id']);
-    }
-
-    /**
-     * @return ActiveQuery
-     */
-    public function getUser()
-    {
-        return $this->hasOne(User::class, ['id' => 'user_id']);
-    }
 
     /**
      * @return array
