+++ models/events/Call.php
@@ -3,8 +3,7 @@
 namespace app\models\events;
 
 use Yii;
-use yii\db\ActiveQuery;
-use yii\db\ActiveRecord;
+use app\models\repositories\Call as CallActiveRecord;
 
 /**
  * This is the model class for table "{{%call}}".
@@ -31,7 +30,7 @@ use yii\db\ActiveRecord;
  * @property Customer $customer
  * @property User $user
  */
-class Call extends ActiveRecord
+class Call extends CallActiveRecord
 {
     const STATUS_NO_ANSWERED = 0;
     const STATUS_ANSWERED = 1;
@@ -41,65 +40,6 @@ class Call extends ActiveRecord
 
     public $duration = 720;
 
-    /**
-     * @inheritdoc
-     */
-    public static function tableName()
-    {
-        return '{{%call}}';
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function rules()
-    {
-        return [
-            [['ins_ts'], 'safe'],
-            [['direction', 'phone_from', 'phone_to', 'type', 'status', 'viewed'], 'required'],
-            [['direction', 'user_id', 'customer_id', 'type', 'status'], 'integer'],
-            [['phone_from', 'phone_to', 'outcome'], 'string', 'max' => 255],
-            [['customer_id'], 'exist', 'skipOnError' => true, 'targetClass' => Customer::class, 'targetAttribute' => ['customer_id' => 'id']],
-            [['user_id'], 'exist', 'skipOnError' => true, 'targetClass' => User::class, 'targetAttribute' => ['user_id' => 'id']],
-        ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function attributeLabels()
-    {
-        return [
-            'id' => Yii::t('app', 'ID'),
-            'ins_ts' => Yii::t('app', 'Date'),
-            'direction' => Yii::t('app', 'Direction'),
-            'directionText' => Yii::t('app', 'Direction'),
-            'user_id' => Yii::t('app', 'User ID'),
-            'customer_id' => Yii::t('app', 'Customer ID'),
-            'status' => Yii::t('app', 'Status'),
-            'statusText' => Yii::t('app', 'Status'),
-            'phone_from' => Yii::t('app', 'Caller Phone'),
-            'phone_to' => Yii::t('app', 'Dialed Phone'),
-            'user.fullname' => Yii::t('app', 'User'),
-            'customer.name' => Yii::t('app', 'Client'),
-        ];
-    }
-
-    /**
-     * @return ActiveQuery
-     */
-    public function getCustomer()
-    {
-        return $this->hasOne(Customer::class, ['id' => 'customer_id']);
-    }
-
-    /**
-     * @return ActiveQuery
-     */
-    public function getUser()
-    {
-        return $this->hasOne(User::class, ['id' => 'user_id']);
-    }
 
     /**
      * @return string
