+++ models/events/User.php
@@ -3,7 +3,7 @@
 namespace app\models\events;
 
 use Yii;
-use yii\db\ActiveRecord;
+use app\models\repositories\User  as UserActiveRecord;
 
 /**
  * This is the model class for table "{{%user}}".
@@ -16,55 +16,9 @@ use yii\db\ActiveRecord;
  * @property integer $updated_at
  * @property string $statusText
  */
-class User extends ActiveRecord
+class User extends UserActiveRecord
 {
-    const STATUS_DELETED = 0;
-    const STATUS_HIDDEN = 1;
-    const STATUS_ACTIVE = 10;
 
-    /**
-     * @inheritdoc
-     */
-    public static function tableName()
-    {
-        return '{{%user}}';
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function rules()
-    {
-        return [
-            [['username', 'created_at', 'updated_at'], 'required'],
-            [[
-                'status',
-                'created_at',
-                'updated_at',
-            ], 'integer'],
-            [[
-                'username',
-                'email',
-            ], 'string', 'max' => 255],
-
-            [['username'], 'unique'],
-
-            ['status', 'default', 'value' => self::STATUS_ACTIVE],
-            ['status', 'in', 'range' => [self::STATUS_ACTIVE, self::STATUS_DELETED, self::STATUS_HIDDEN]],
-        ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function attributeLabels()
-    {
-        return [
-            'id' => Yii::t('app', 'ID'),
-            'username' => Yii::t('app', 'Username (login)'),
-            'statusText' => Yii::t('app', 'Status'),
-        ];
-    }
 
     /**
      * @return array
