+++ models/repositories/Customer.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace app\models\repositories;
+
+
+use Yii;
+use yii\db\ActiveRecord;
+
+
+/**
+ * This is the model class for table "{{%customer}}".
+ *
+ * @property integer $id
+ * @property string $name
+ */
+class Customer extends ActiveRecord
+{
+
+    /**
+     * @inheritdoc
+     */
+    public static function tableName()
+    {
+        return '{{%customer}}';
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function rules()
+    {
+        return [
+            [['name'], 'string', 'max' => 255],
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function attributeLabels()
+    {
+        return [
+            'name' => Yii::t('app', 'Name'),
+        ];
+    }
+
+}
